test(tiers): move Brain Monkey tier-gating tests from Integration to Unit
The three tier-gating tests (Generator, Images, Seo) used Brain Monkey /
Patchwork to mock WordPress functions. Running them under the integration
bootstrap (real WordPress already loaded) caused
Patchwork\Exceptions\DefinedTooEarly when trying to intercept
register_rest_route(), which is already defined by that point.

A secondary issue: their tearDown() replaced $wpdb with a WpdbStubFactory
stub that lacked db_connect(), breaking subsequent real integration tests
(SeoGenerateTest, TierFeatureMatrixTest) that rely on the actual $wpdb.

Fix: move all three files to tests/Unit/Modules/*/, update namespaces,
add get_option mock to setUp so TierManager::get_user_tier() takes the
user-meta path, add stilus_trial_started to the trial test mock so
is_trial_active() returns true, and remove the $wpdb stub tearDown.
