+++ tests/Unit/Modules/Seo/SeoTierGatingTest.php
@@ -1,21 +1,20 @@
 <?php
 declare( strict_types=1 );
 
-namespace Stilus\Tests\Integration\Generator;
+namespace Stilus\Tests\Unit\Modules\Seo;
 
 use Brain\Monkey;
 use Brain\Monkey\Functions;
-use Stilus\Modules\Generator\GeneratorModule;
-use Stilus\Tests\Helpers\WpdbStubFactory;
+use Stilus\Modules\Seo\SeoModule;
 use PHPUnit\Framework\TestCase;
 
 /**
- * Tier-gating integration tests for the Generator module REST route.
+ * Tier-gating unit tests for the SEO module REST route.
  *
- * Exercises the permission_callback chain — TierManager::user_can('generator')
+ * Exercises the permission_callback chain — TierManager::user_can('seo')
  * returns false for the free tier and true for trial+.
  */
-class GeneratorTierGatingTest extends TestCase {
+class SeoTierGatingTest extends TestCase {
 
 	/** @var array<string, array<string, mixed>> */
 	private array $captured_routes = [];
@@ -23,6 +22,10 @@ class GeneratorTierGatingTest extends TestCase {
 	protected function setUp(): void {
 		parent::setUp();
 		Monkey\setUp();
+
+		// Ensure TierManager::get_user_tier() sees 'free' site tier so the user-meta
+		// path is exercised rather than the pro-site short-circuit.
+		Functions\when( 'get_option' )->justReturn( 'free' );
 
 		// Capture the registered routes so we can invoke permission_callback directly.
 		$this->captured_routes = [];
@@ -32,23 +35,21 @@ class GeneratorTierGatingTest extends TestCase {
 			}
 		);
 
-		GeneratorModule::register_routes();
+		SeoModule::register_routes();
 	}
 
 	protected function tearDown(): void {
-		global $wpdb;
-		$wpdb = WpdbStubFactory::create(); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Intentional test stub.
 		Monkey\tearDown();
 		parent::tearDown();
 	}
 
-	public function test_generator_generate_returns_403_for_free_tier_user(): void {
-		$this->assertArrayHasKey( '/generate', $this->captured_routes );
-		$permission_callback = $this->captured_routes['/generate']['permission_callback'];
+	public function test_seo_generate_returns_403_for_free_tier_user(): void {
+		$this->assertArrayHasKey( '/seo/generate', $this->captured_routes );
+		$permission_callback = $this->captured_routes['/seo/generate']['permission_callback'];
 
 		$month_key = 'stilus_usage_' . gmdate( 'Y_m' );
 
-		// Free tier: edit_posts capability present, within usage limit, but generator feature is disabled.
+		// Free tier: edit_posts capability present, within usage limit, but seo feature is disabled.
 		Functions\when( 'current_user_can' )->justReturn( true );
 		Functions\when( 'get_current_user_id' )->justReturn( 1 );
 		Functions\when( 'get_user_meta' )->alias(
@@ -63,23 +64,26 @@ class GeneratorTierGatingTest extends TestCase {
 			}
 		);
 
-		// permission_callback returns false because TierConfig::FEATURES['free']['generator'] = false.
+		// permission_callback returns false because TierConfig::FEATURES['free']['seo'] = false.
 		$this->assertFalse( (bool) $permission_callback() );
 	}
 
-	public function test_generator_generate_returns_200_for_trial_tier_user(): void {
-		$this->assertArrayHasKey( '/generate', $this->captured_routes );
-		$permission_callback = $this->captured_routes['/generate']['permission_callback'];
+	public function test_seo_generate_returns_200_for_trial_tier_user(): void {
+		$this->assertArrayHasKey( '/seo/generate', $this->captured_routes );
+		$permission_callback = $this->captured_routes['/seo/generate']['permission_callback'];
 
 		$month_key = 'stilus_usage_' . gmdate( 'Y_m' );
 
-		// Trial tier: edit_posts capability present, within usage limit, generator feature enabled.
+		// Trial tier: edit_posts capability present, within usage limit, seo feature enabled.
 		Functions\when( 'current_user_can' )->justReturn( true );
 		Functions\when( 'get_current_user_id' )->justReturn( 2 );
 		Functions\when( 'get_user_meta' )->alias(
 			function ( $user_id, $key, $single = false ) use ( $month_key ) {
 				if ( 'stilus_tier' === $key ) {
 					return 'trial';
+				}
+				if ( 'stilus_trial_started' === $key ) {
+					return (string) time(); // trial started now, well within the trial period
 				}
 				if ( $month_key === $key ) {
 					return '0'; // well within 300k trial limit
@@ -88,7 +92,7 @@ class GeneratorTierGatingTest extends TestCase {
 			}
 		);
 
-		// permission_callback returns true because TierConfig::FEATURES['trial']['generator'] = true.
+		// permission_callback returns true because TierConfig::FEATURES['trial']['seo'] = true.
 		$this->assertTrue( (bool) $permission_callback() );
 	}
 }
